app/ai/file_parser.py
from PyPDF2 import PdfReader
import docx

def read_pdf(file):
    reader = PdfReader(file)
    text = ""
    for page in reader.pages:
        text += page.extract_text() or ""
    return text

def read_docx(file):
    document = docx.Document(file)
    text = "\n".join([para.text for para in document.paragraphs])
    return text

def read_txt(file):
    return file.read().decode("utf-8")
