requirements.txt
fastapi==0.128.64
uvicorn==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.22
pydantic==2.12.5
spacy==3.8.11
scikit-learn==1.8.0
PyPDF2==3.0.1
python-docx==1.2.0
requests==2.32.5
